requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# YouTube Download
yt-dlp>=2023.10.7

# Video Processing
moviepy>=1.0.3
opencv-python>=4.8.1.78
Pillow>=10.1.0

# Audio/Speech Processing
openai-whisper>=20231117
pydub>=0.25.1

# NLP & ML
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2
nltk>=3.8.1
spacy>=3.7.2
scikit-learn>=1.3.2

# Dataset Processing
pandas>=2.1.3
numpy>=1.26.2

# Utilities
tqdm>=4.66.1
requests>=2.31.0

